Fix signing of rpms with a size greater than 4GByte
The main fix is changing lensig in rpm_readheaderpayload to be
64bit instead of 32bit. With 32bit, it will wrap around to zero
after the first 4GByte are read. This triggers a re-calculation
of the header size, leading to a misleading "bad header size"
error.

Also prepare to read the 64bit size if present.

+++ inc.h
@@ -140,7 +140,7 @@ struct rpmdata {
   byte *rpmsig;         /* signature data (with room for two new signatures) */
 
   u32 rpmdataoff;
-  u32 hdrin_size;
+  unsigned long long hdrin_size;
   byte *hdrin_md5;      /* points into rpmsig */
   int gotsha1;          /* did we see a RPMSIGTAG_SHA1 tag? */
